requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==24.3.0
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
fonttools==4.55.4
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.28.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.5
kaggle==1.6.17
kiwisolver==1.4.8
lightly==1.5.17
lightly-utils==0.0.2
lightning-utilities==0.11.9
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.2
setuptools==75.1.0
six==1.17.0
sympy==1.13.1
text-unidecode==1.3
timm==1.0.14
torch==2.4.0+cu121
torchmetrics==1.6.1
torchvision==0.19.0+cu121
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
wheel==0.44.0
xformers==0.0.27.post2
yarl==1.18.3
